// src/__generated__/LikeFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: LikeFragment
// ====================================================

export interface LikeFragment {
  __typename: "Photo";
  isLiked: boolean;
  likes: number;
}